backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0
auth0-server-python>=0.1.0
httpx>=0.27.0
google-api-python-client>=2.150.0
google-auth>=2.28.0
google-auth-httplib2>=0.2.0
twilio>=9.0.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
groq>=0.5.0
